# Reduce memory allocations in add function